# backend/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_modules_identity_internal_transport.AssignRolesReq:
    properties:
      role_ids:
        items:
          type: integer
        maxItems: 10
        type: array
    required:
    - role_ids
    type: object
...
  internal_modules_identity_internal_transport.CreateUserReq:
    properties:
      email:
        maxLength: 255
        type: string
      name:
        maxLength: 255
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      password_confirmation:
        type: string
      role_ids:
        items:
          type: integer
        maxItems: 10
        type: array
    required:
    - email
    - name
    - password
    type: object
...
  internal_modules_identity_internal_transport.LoginReq:
    properties:
      email:
        maxLength: 255
        type: string
      password:
        maxLength: 4096
        type: string
    required:
    - email
    - password
    type: object
  internal_modules_identity_internal_transport.MeResp:
    properties:
      tenant:
        $ref: '#/definitions/internal_modules_identity_internal_transport.TenantResp'
      user:
        $ref: '#/definitions/internal_modules_identity_internal_transport.UserResp'
    type: object
  internal_modules_identity_internal_transport.PlatformUpdateTenantReq:
    properties:
      active:
        type: boolean
    required:
    - active
    type: object
  internal_modules_identity_internal_transport.RefreshReq:
    properties:
      refresh_token:
        maxLength: 1024
        type: string
    required:
    - refresh_token
    type: object
  internal_modules_identity_internal_transport.RegisterReq:
    properties:
      email:
        maxLength: 255
        type: string
      name:
        maxLength: 255
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      password_confirmation:
        description: PasswordConfirmation is optional; when sent it must match.
        type: string
      tenant_name:
        maxLength: 255
        type: string
      timezone:
        maxLength: 64
        type: string
    required:
    - email
    - password
    - tenant_name
    type: object
  internal_modules_identity_internal_transport.RoleResp:
    properties:
      id:
        type: integer
      key:
        type: string
      label:
        type: string
    type: object
  internal_modules_identity_internal_transport.SessionResp:
    properties:
      access_token:
        type: string
      expires_in:
        description: seconds
        type: integer
      refresh_token:
        type: string
      tenant:
        $ref: '#/definitions/internal_modules_identity_internal_transport.TenantResp'
      token_type:
        type: string
      user:
        $ref: '#/definitions/internal_modules_identity_internal_transport.UserResp'
    type: object
  internal_modules_identity_internal_transport.TenantListResp:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_modules_identity_internal_transport.TenantResp'
        type: array
      next_cursor:
        type: string
    type: object
  internal_modules_identity_internal_transport.TenantResp:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      owner_user_id:
        type: string
      settings:
        additionalProperties: {}
        type: object
      slug:
        type: string
      timezone:
        type: string
      updated_at:
        type: string
    type: object
  internal_modules_identity_internal_transport.UpdateTenantReq:
    properties:
      name:
        maxLength: 255
        minLength: 1
        type: string
      settings:
        additionalProperties: {}
        type: object
      timezone:
        maxLength: 64
        minLength: 1
        type: string
    type: object
  internal_modules_identity_internal_transport.UpdateUserReq:
    properties:
      active:
        type: boolean
      current_password:
        maxLength: 4096
        type: string
      email:
        maxLength: 255
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
    type: object
  internal_modules_identity_internal_transport.UserListResp:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_modules_identity_internal_transport.UserResp'
        type: array
      next_cursor:
        type: string
    type: object
  internal_modules_identity_internal_transport.UserResp:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      email:
        type: string
      email_verified_at:
        type: string
      id:
        type: string
      name:
        type: string
      role_ids:
        items:
          type: integer
        type: array
      roles:
        items:
          $ref: '#/definitions/internal_modules_identity_internal_transport.RoleResp'
        type: array
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
...
  levelup_internal_platform_httpx.Problem:
    properties:
      code:
        type: string
      detail:
        type: string
      errors:
        additionalProperties:
          type: string
        type: object
      status:
        type: integer
      title:
        type: string
      trace_id:
        type: string
      type:
        type: string
    type: object
info:
  contact: {}
  description: Modular monolith blueprint API.
  title: LevelUp API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_internal_transport.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_identity_internal_transport.SessionResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
      summary: Log in with email and password
      tags:
      - auth
  /auth/logout:
    post:
      description: Deny-lists the presented access token and revokes every refresh
        token of the user.
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
      security:
      - BearerAuth: []
      summary: Log out everywhere
      tags:
      - auth
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_identity_internal_transport.MeResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
      security:
      - BearerAuth: []
      summary: The authenticated user, roles and tenant
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Consumes the refresh token and returns a new pair; a replayed token
        revokes every session of the user.
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_internal_transport.RefreshReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_identity_internal_transport.SessionResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
      summary: Rotate a refresh token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Atomically creates the tenant, the owner user and the owner role,
        and returns a session.
      parameters:
      - description: Registration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_internal_transport.RegisterReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_modules_identity_internal_transport.SessionResp'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
      summary: Register a tenant and its owner
      tags:
      - auth
...
  /platform/tenants:
    get:
      parameters:
      - description: Page size (default 25, max 100)
        in: query
        name: limit
        type: integer
      - description: Opaque cursor from the previous page
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_identity_internal_transport.TenantListResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
      security:
      - BearerAuth: []
      summary: List tenants (platform admins)
      tags:
      - platform
  /platform/tenants/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Tenant id
        in: path
        name: id
        required: true
        type: string
      - description: Active flag
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_internal_transport.PlatformUpdateTenantReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_identity_internal_transport.TenantResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
      security:
      - BearerAuth: []
      summary: Activate or deactivate a tenant (platform admins)
      tags:
      - platform
...
  /tenant:
    delete:
      description: Soft delete; every module purges its data on tenant.deleted.v1.
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
      security:
      - BearerAuth: []
      summary: Delete the caller's tenant (owner only)
      tags:
      - tenant
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_identity_internal_transport.TenantResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
      security:
      - BearerAuth: []
      summary: The caller's tenant
      tags:
      - tenant
    patch:
      consumes:
      - application/json
      parameters:
      - description: Fields to change; settings replaces the whole object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_internal_transport.UpdateTenantReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_identity_internal_transport.TenantResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
      security:
      - BearerAuth: []
      summary: Update the caller's tenant (partial)
      tags:
      - tenant
  /users:
    get:
      parameters:
      - description: Page size (default 25, max 100)
        in: query
        name: limit
        type: integer
      - description: Opaque cursor from the previous page
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_identity_internal_transport.UserListResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
      security:
      - BearerAuth: []
      summary: List the users of the caller's tenant
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: User
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_internal_transport.CreateUserReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_modules_identity_internal_transport.UserResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
      security:
      - BearerAuth: []
      summary: Create a user in the caller's tenant
      tags:
      - users
  /users/{id}:
    delete:
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
      security:
      - BearerAuth: []
      summary: Delete a user of the caller's tenant
      tags:
      - users
    get:
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_identity_internal_transport.UserResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
      security:
      - BearerAuth: []
      summary: Get a user of the caller's tenant
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Self-service is always allowed; changing your own password requires
        current_password.
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      - description: Fields to change
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_internal_transport.UpdateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_identity_internal_transport.UserResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
      security:
      - BearerAuth: []
      summary: Update a user (partial)
      tags:
      - users
  /users/{id}/roles:
    put:
      consumes:
      - application/json
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      - description: Role ids
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_internal_transport.AssignRolesReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_identity_internal_transport.UserResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/levelup_internal_platform_httpx.Problem'
      security:
      - BearerAuth: []
      summary: Replace a user's roles
      tags:
      - users
...
swagger: "2.0"